Initial Version of the webapp
Created some views in views.py, handled the url patterns to map those views to render me the actual webpage. CSS files and other related things are stored in the static directory.
created some models to work with databases. Created some forms to include them in my webpages and also to test the power of django.